chore: clean up repo, remove old files
