--- src/@new/constants/index.ts
@@ -1,3 +1,0 @@
-export * from './abi/marketplace';
-export * from './abi/swivel';
-export * from './eip712';
