--- src/@new/index.ts
@@ -1,18 +1,0 @@
-// import { EthersVendor } from './vendors/ethers/ethers';
-// import * as EthersUtils from './vendors/ethers/utils';
-// import { Web3Vendor } from './vendors/web3/web3';
-
-export * from './constants';
-export * from './interfaces';
-export * from './marketplace';
-export * from './swivel';
-export * from './vendors';
-
-// export const ethers = {
-//     Vendor: EthersVendor,
-//     utils: EthersUtils,
-// };
-
-// export const web3 = {
-//     Vendor: Web3Vendor,
-// };
