--- src/@new/interfaces/index.ts
@@ -1,7 +1,0 @@
-export * from './abi';
-export * from './contracts';
-export * from './eip712';
-export * from './order';
-export * from './signature';
-export * from './transaction';
-export * from './vendor';
