--- src/@new/interfaces/order.ts
@@ -1,11 +1,0 @@
-export interface Order {
-    key: string;
-    maker: string;
-    underlying: string;
-    vault: boolean;
-    exit: boolean;
-    principal: string;
-    premium: string;
-    maturity: string;
-    expiry: string;
-}
