--- src/@new/interfaces/signature.ts
@@ -1,5 +1,0 @@
-export interface Signature {
-    v: number | string;
-    r: string;
-    s: string;
-}
