--- src/@new/interfaces/transaction.ts
@@ -1,12 +1,0 @@
-export interface TxOptions {
-    to?: string;
-    from?: string;
-    gasPrice?: number | string;
-    gasLimit?: number | string;
-    value?: number | string;
-    data?: unknown;
-}
-
-export interface TxResponse {
-    hash: string;
-}
