--- src/@types/index.ts
@@ -1,2 +1,0 @@
-export type Nos = number | string;
-export type Abi = Array<any>;
