--- src/abstracts/deployed.ts
@@ -1,40 +1,0 @@
-/**
- * @remarks
- * Base class of all Higher Order Contracts (HOCs).
- * Note that we do not want to couple this class to any specific Vendor.
- */
-
-import { Abi } from '../@types';
-import { Keyed, Contract, TxOpts } from '../interfaces';
-import Vendor from './vendor';
-
-export default abstract class implements Keyed {
-    [key: string]: any
-
-    public abi: Abi;
-
-    public vendor: Vendor;
-
-    public contract?: Contract;
-
-    /**
-   * @param v - Optional Vendor the HOC will use
-   * @param a - Compiled abi of the smart contract this HOC wraps
-   */
-    protected constructor (v: Vendor, a: Abi) {
-        this.vendor = v;
-        this.abi = a;
-    }
-
-    /**
-   * @remarks Hoists the vendor specific contract wrapping to `this.contract`
-   * @param a - ETH address of an already deployed smart contract
-   * @param o - Optional specified transaction options
-   *
-   * @returns boolean indicating a successful wrapping of the target deployed contract
-   */
-    at (a: string, o?: TxOpts): boolean {
-        this.contract = this.vendor.contract(a, this.abi, o);
-        return !!this.contract;
-    }
-}
