+++ src/errors/index.ts
@@ -1,13 +1,7 @@
-/**
- * @remarks
- * It could be argued that errors be placed in ../constants.
- * This may be true, however, it is undetermined if we would decide to
- * instantiate actual instances of errors, or write custom error constructors,
- * therefore, we will keep them in their own dir.
- */
+export const CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED = 'chain-id and verifying contract required';
 
-export * from './vendor';
+export const CONTRACT_INSTANTIATION_FAILED = (contract: 'marketplace' | 'swivel', address: string): string =>
+    `Failed to instantiate deployed ${ contract } contract at ${ address }.`;
 
-export const OWNER_REQUIRED = 'owner required';
-export const PROVIDER_REQUIRED = 'provider required';
-export const CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED = 'chain id and verifying contract required';
+export const MISSING_CONTRACT_ADDRESS = (contract: 'marketplace' | 'swivel'): string =>
+    `Missing contract address for deployed ${ contract } contract. Set an address using ${ contract }.at(<contract_address>).`;
