--- src/errors/vendor.ts
@@ -1,1 +1,0 @@
-export const SIGNER_OR_PROVIDER_REQUIRED = 'signer or provider required';
