+++ src/index.ts
@@ -1,9 +1,5 @@
-import Swivel from './swivel';
-import Vendor from './vendors/ethers';
-import * as ethersHelper from './vendors/ethers.helpers';
-
-const helpers = {
-    ethers: ethersHelper,
-};
-
-export { Swivel, Vendor, helpers };
+export * from './constants';
+export * from './interfaces';
+export * from './marketplace';
+export * from './swivel';
+export * from './vendors';
