+++ src/interfaces/eip712.ts
@@ -1,10 +1,6 @@
 // EIP-712 Typed Data
 // See: https://eips.ethereum.org/EIPS/eip-712
 
-// export type Bytes = ArrayLike<number>;
-// export type BytesLike = Bytes | string;
-// export type BigNumberLike = bigint | string | number;
-
 export interface TypedDataDomain {
     name?: string;
     version?: string;
