+++ src/interfaces/index.ts
@@ -1,5 +1,7 @@
-export * from './contract';
-export * from './keyed';
+export * from './abi';
+export * from './contracts';
+export * from './eip712';
 export * from './order';
 export * from './signature';
 export * from './transaction';
+export * from './vendor';
