--- src/interfaces/keyed.ts
@@ -1,3 +1,0 @@
-export interface Keyed {
-    [key: string]: any;
-}
