+++ src/interfaces/vendor.ts
