+++ src/swivel.ts
@@ -1,50 +1,97 @@
-import Deployed from './abstracts/deployed';
-import Vendor from './abstracts/vendor';
 import { SWIVEL_ABI } from './constants';
-import { CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED } from './errors';
-import { Components, Order, TxResponse } from './interfaces';
+import { CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED, CONTRACT_INSTANTIATION_FAILED, MISSING_CONTRACT_ADDRESS } from './errors';
+import { Order, SwivelContract, TxOptions, TxResponse, Vendor } from './interfaces';
 
-export default class extends Deployed {
-    public chainId: any;
+export class Swivel implements SwivelContract {
 
-    public verifyingContract: any;
+    protected contract?: SwivelContract;
+
+    protected options?: TxOptions;
+
+    protected abi = SWIVEL_ABI;
+
+    address?: string;
+
+    vendor: Vendor;
+
+    chainId?: number;
+
+    verifyingContract?: string;
 
     /**
-     * @param vendor - Instance of a Vendor class
-     * @param i - optional chainId for the deployed smart contract. NOTE: signOrder requires this be set
-     * @param verifier - optional address of a deployed verifying contract. NOTE: signOrder requires this be set
+     * Create a new swivel instance
+     *
+     * @param v - a vendor instance (ethers.js or web3.js vendor)
+     * @param i - optional chain-id for the deployed smart contract; NOTE: signOrder requires this be set
+     * @param c - optional address of a deployed verifying contract; NOTE: signOrder requires this be set
      */
-    constructor (vendor: Vendor, i?: number, verifier?: string) {
-        super(vendor, SWIVEL_ABI);
+    constructor (v: Vendor, i?: number, c?: string) {
+
+        this.vendor = v;
+
+        // TODO: what is the verifying contract?
         this.chainId = i;
-        this.verifyingContract = verifier;
+        this.verifyingContract = c;
     }
 
     /**
-     * @remarks
-     * Proxy method which delegates to its vendor's signOrder method.
+     * Set the deployed smart contract address
      *
-     * @param o - raw order object
-     * @return Vendor specific Promise
+     * @remarks
+     * Creates a vendor specific instance of the deployed swivel smart contract
+     * and wraps it in a generic contract representation.
+     *
+     * @example
+     * // TODO: make a good example
+     * ```
+     * const ethersProvider = getDefaultProvider();
+     * const signer = Wallet.createRandom().connect(ethersProvider);
+     * const vendor = new EthersVendor(ethersProvider, signer);
+     * const swivel = new Swivel(vendor, 1, '0x123').at('0xabc');
+     * ```
+     *
+     * @param a - ETH address of the already deployed swivel smart contract
+     * @param o - optional transaction options
+     *
+     * @returns the swivel instance for chaining
      */
+    at (a: string, o?: TxOptions): Swivel {
+
+        this.contract = this.vendor.contracts.swivel(a, this.abi, o);
+
+        if (!this.contract) throw CONTRACT_INSTANTIATION_FAILED('swivel', a);
+
+        this.address = this.contract.address;
+        this.options = o;
+
+        return this;
+    }
+
     async signOrder (o: Order): Promise<string> {
-        if (!this.chainId || !this.verifyingContract) return Promise.reject(CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED);
+
+        if (!this.chainId || !this.verifyingContract) throw CHAIN_ID_AND_VERIFYING_CONTRACT_REQUIRED;
+
         return await this.vendor.signOrder(o, this.chainId, this.verifyingContract);
     }
 
-    /**
-     * @remarks
-     * Calls to its vendor to prepare the order, gets the signature components then attempts to cancel
-     * via the swivel contract method
-     *
-     * @param o - order
-     * @param s - signature
-     * @return Promise resolving to a transaction response
-     */
-    async cancel (o: Order, s: string): Promise<TxResponse> {
-        const order = this.vendor.prepareOrder(o);
-        const components: Components = this.vendor.splitSignature(s);
+    async initiate (o: Order[], a: number[], s: string[]): Promise<TxResponse> {
 
-        return await this.contract?.functions.cancel(order, components) as Promise<TxResponse>;
+        if (!this.contract) throw MISSING_CONTRACT_ADDRESS('swivel');
+
+        return await this.contract.initiate(o, a, s);
+    }
+
+    async exit (o: Order[], a: number[], s: string[]): Promise<TxResponse> {
+
+        if (!this.contract) throw MISSING_CONTRACT_ADDRESS('swivel');
+
+        return await this.contract.exit(o, a, s);
+    }
+
+    async cancel (o: Order, s: string): Promise<TxResponse> {
+
+        if (!this.contract) throw MISSING_CONTRACT_ADDRESS('swivel');
+
+        return await this.contract.cancel(o, s);
     }
 }
