+++ src/vendors/ethers/utils/prepare-amount.ts
@@ -1,11 +1,11 @@
-import { BigNumber, ethers } from 'ethers';
+import { ethers } from 'ethers';
 
 /**
  * Converts an amount into an `ethers.js` specific amount.
  *
  * @param a - amount
  */
-export function prepareAmount (a: number | string): BigNumber {
+export function prepareAmount (a: number | string): ethers.BigNumber {
 
     return ethers.BigNumber.from(a);
 }
