+++ src/vendors/index.ts
