--- src/vendors/web3.ts
@@ -1,1 +1,0 @@
-// TODO if needed
